move `WidgetbookTheme` to theme addon
