+++ packages/widgetbook/lib/src/addons/theme_addon/addon.dart
@@ -2,3 +2,4 @@ export 'cupertino_theme_addon.dart';
 export 'material_theme_addon.dart';
 export 'theme_addon.dart';
 export 'theme_setting.dart';
+export 'widgetbook_theme.dart';
