+++ packages/widgetbook/lib/src/addons/theme_addon/material_theme_addon.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-import '../../theming/widgetbook_theme.dart';
 import 'theme_addon.dart';
+import 'widgetbook_theme.dart';
 
 class MaterialThemeAddon extends ThemeAddon<ThemeData> {
   MaterialThemeAddon({
