+++ packages/widgetbook/lib/widgetbook.dart
@@ -1,5 +1,4 @@
 export 'src/addons/addons.dart';
 export 'src/knobs/knobs_builder.dart';
 export 'src/navigation/navigation.dart';
-export 'src/theming/widgetbook_theme.dart';
 export 'src/widgetbook.dart';
